Fixed invalid Offers from throwing when reading
closes https://github.com/TryGhost/Product/issues/3675
refs https://github.com/TryGhost/Ghost/commit/cf81daa6da0a60898e463b5774b5f1760767b444

As part of our architecture guidelines Repository implementations should protect
against invalid or malformed data in persistence. We do not want read operations
of Entities to throw because of such data. For some fields that bad data can be
fixed or handled in the constructor or static create factory method and replaced
with valid data, others will cause the factory to throw.

This means that Repositories should catch these errors and exclude those
entities from their results. We log the errors in Sentry so that we have
visibility on the state of bad data in DBs