Implemented email customisation labs flag in AdminX (#17881)
refs https://github.com/TryGhost/Product/issues/3745

---

This pull request adds a new feature to the `admin-x-settings` app that
allows users to customize the colors of their email newsletters. It also
simplifies and refactors the routing logic and the components related to
integrations, newsletters, and users. It introduces new dependencies and
components for color manipulation and selection, and updates the
existing components and stories to use them.