🐛 Fixed admin crashing when deleting a tier benefit that was last moved (#20628)
ref https://linear.app/tryghost/issue/ENG-1433/
- in the tiers details modal, if a tier was deleted after being moved it
would crash admin

The SortableList component calls renderItem for the drag overlay on the
dragged item. However, if that item was deleted, the draggingId never
got reset causing the drag overlay to try to call renderItem on a
nonexistent item.

As a note, we don't actually use this overlay... it could be best to
strip it out to prevent causing issues.